vamp_setup/VAMP-0.3.3.1/requirements.txt
Cython>=0.17.4
argparse>=1.2.1
numpy>=1.6.2
biopython>=1.57
bx-python>=0.7.1
cutadapt>=1.2.1
distribute>=0.6.34
paired-sequence-utils>=0.1
pybedtools==0.6
wsgiref>=0.1.2
vamp==%__VERSION__%